requirements.txt
charset-normalizer==3.4.5
colorama==0.4.6
ExifRead==3.5.1
pillow==12.1.1
python-magic==0.4.27
reportlab==4.4.10
tabulate==0.10.0
